core: Improve separation of concerns
- decouple thermostats from long-range GPU methods
- narrow includes of thermostat.hpp, cells.hpp, rotation.hpp
- break cyclic dependency in bonded interactions
- move functions to different files as necessary
- pass cell_structure global as function argument as necessary
